# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Situation X - AI + Astrology Situation Analysis API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feedback:
    post:
      operationId: createFeedback
      tags: [feedback]
      summary: Submit feedback for an analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackRequest"
      responses:
        "200":
          description: Feedback saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackItem"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getFeedbackList
      tags: [feedback]
      summary: Get all feedback (optionally filtered by analysis)
      parameters:
        - name: analysisId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Feedback list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackListResponse"
  /feedback/{id}:
    delete:
      operationId: deleteFeedback
      tags: [feedback]
      summary: Delete feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    CreateFeedbackRequest:
      type: object
      properties:
        analysisId:
          type: integer
          description: ID of the analysis being reviewed
        rating:
          type: integer
          description: 1-5 star rating
          minimum: 1
          maximum: 5
        accuracy:
          type: integer
          description: How accurate was the analysis? 1-5
          minimum: 1
          maximum: 5
        comment:
          type: string
          description: Free text feedback
        helpful:
          type: boolean
          description: Was this analysis helpful?
        outcome:
          type: string
          enum: [matched, partly, different]
          description: Follow-up - did things turn out the way the reading suggested? (older clients; the server derives it from actionTaken and result when those are sent)
        actionTaken:
          type: string
          enum: [followed, other, nothing]
          description: What the person did - the suggested step, another step, or nothing
        result:
          type: string
          enum: [better, same, worse]
          description: How things actually turned out
        reasonTags:
          type: array
          maxItems: 9
          items:
            type: string
            enum: [advice_right, missed_perspective, missed_facts, too_generic, timing_off, astrology_helped, astrology_off, steps_helpful, steps_unrealistic]
          description: What was right or missed
      required:
        - analysisId
        - rating
    FeedbackItem:
      type: object
      properties:
        id:
          type: integer
        analysisId:
          type: integer
        situationSnippet:
          type: string
        rating:
          type: integer
        accuracy:
          type: integer
        comment:
          type: string
        helpful:
          type: boolean
        outcome:
          type: string
        actionTaken:
          type: string
        result:
          type: string
        reasonTags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - analysisId
        - rating
        - createdAt
    FeedbackListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FeedbackItem"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message