# lib/api-zod/src/generated/types/createFeedbackRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Situation X - AI + Astrology Situation Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFeedbackRequestActionTaken } from "./createFeedbackRequestActionTaken";
import type { CreateFeedbackRequestOutcome } from "./createFeedbackRequestOutcome";
import type { CreateFeedbackRequestReasonTagsItem } from "./createFeedbackRequestReasonTagsItem";
import type { CreateFeedbackRequestResult } from "./createFeedbackRequestResult";

export interface CreateFeedbackRequest {
  /** ID of the analysis being reviewed */
  analysisId: number;
  /**
   * 1-5 star rating
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /**
   * How accurate was the analysis? 1-5
   * @minimum 1
   * @maximum 5
   */
  accuracy?: number;
  /** Free text feedback */
  comment?: string;
  /** Was this analysis helpful? */
  helpful?: boolean;
  /** Follow-up - did things turn out the way the reading suggested? (older clients; the server derives it from actionTaken and result when those are sent) */
  outcome?: CreateFeedbackRequestOutcome;
  /** What the person did - the suggested step, another step, or nothing */
  actionTaken?: CreateFeedbackRequestActionTaken;
  /** How things actually turned out */
  result?: CreateFeedbackRequestResult;
  /**
   * What was right or missed
   * @maxItems 9
   */
  reasonTags?: CreateFeedbackRequestReasonTagsItem[];
}